Address stitched audio review feedback
